Add file header block and more completions.
Adding a snippet at the beginning of the file will add a file header comment with an author and copyright block that can be specified in the settings file.
A section line length can be specified to create a consistant boundary around certain sections.

Many more completions are also added.